needed to remove OutputDirectoryLogging.catchLogEntries(); from RunBerlinScenario
